requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai>=0.8.3
numpy
pillow
requests